requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
matplotlib>=3.8.0
pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
reportlab>=4.0.0
kaleido>=0.2.1
selenium>=4.13.0
webdriver-manager>=4.0.0
mplsoccer>=1.5.0
scipy>=1.11.0
scikit-learn>=1.3.0
